factored for unit tests
